Update GitHub Actions workflow for CI/CD
Renamed workflow to ".NET CI/CD" and job to "build-and-test".
Added new "release" job dependent on "build-and-test".
The "release" job runs on the "main" branch and includes steps
for checking out code, setting up .NET, restoring dependencies,
building, packing, and pushing the package to NuGet if the
version has changed.